Extract unpacking functionality to a struct
